refactor(ui): extract VirtualQuotaFallbackProviderPresentation
Extracts the presentation logic for `VirtualQuotaFallbackProvider` into a new `VirtualQuotaFallbackProviderPresentation` component. This separates the UI rendering from the component's state management and logic, improving maintainability and testability.

The `VirtualQuotaFallbackProvider` now handles state, data fetching, and event handlers, passing necessary props to the new presentation component. The `VirtualLimitInputs` component has also been renamed to `VirtualLimitInputsPresentation` and moved into the new presentation file.

This change also updates the Storybook story to use the new presentation component and adjusts the `LabeledProgress` component's styling for better visual integration.
