+++ webview-ui/src/components/settings/providers/VirtualQuotaFallbackProvider.tsx
@@ -1,24 +1,8 @@
 import { useCallback, useEffect, useMemo, useState } from "react"
-import { Trans, useTranslation } from "react-i18next"
-import { VSCodeButton, VSCodeTextField } from "@vscode/webview-ui-toolkit/react"
-import { PlusIcon, TrashIcon } from "@radix-ui/react-icons"
-import { ChevronUp, ChevronDown } from "lucide-react"
-
 import { type ProviderSettings, type ProviderSettingsEntry } from "@roo-code/types"
 import { vscode } from "@src/utils/vscode"
 import { useExtensionState } from "@src/context/ExtensionStateContext"
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue, LabeledProgress } from "@src/components/ui"
-import {
-	AlertDialog,
-	AlertDialogAction,
-	AlertDialogCancel,
-	AlertDialogContent,
-	AlertDialogDescription,
-	AlertDialogFooter,
-	AlertDialogHeader,
-	AlertDialogTitle,
-} from "@src/components/ui/alert-dialog"
-import { inputEventTransform } from "../transforms"
+import { VirtualQuotaFallbackProviderPresentation } from "./VirtualQuotaFallbackProviderPresentation"
 import { UsageResultByDuration } from "@roo-code/types"
 
 type VirtualQuotaFallbackProviderProps = {
@@ -26,7 +10,7 @@ type VirtualQuotaFallbackProviderProps = {
 	setApiConfigurationField: (field: keyof ProviderSettings, value: ProviderSettings[keyof ProviderSettings]) => void
 }
 
-type VirtualQuotaFallbackProviderData = {
+export type VirtualQuotaFallbackProviderData = {
 	profileName?: string
 	profileId?: string
 	profileLimits?: {
@@ -39,19 +23,13 @@ type VirtualQuotaFallbackProviderData = {
 	}
 }
 
-type LimitInputsProps = {
-	profile: VirtualQuotaFallbackProviderData
-	index: number
-	onProfileChange: (index: number, profile: VirtualQuotaFallbackProviderData) => void
-}
-
 export const VirtualQuotaFallbackProvider = ({
 	apiConfiguration,
 	setApiConfigurationField,
 }: VirtualQuotaFallbackProviderProps) => {
 	const { listApiConfigMeta, currentApiConfigName } = useExtensionState()
 	const [isAlertOpen, setIsAlertOpen] = useState(false)
-	const { t } = useTranslation()
+	const [usageData, setUsageData] = useState<Record<string, UsageResultByDuration>>({})
 
 	// Get current profile ID to exclude from available profiles
 	const currentProfile = listApiConfigMeta?.find((config) => config.name === currentApiConfigName)
@@ -115,6 +93,7 @@ export const VirtualQuotaFallbackProvider = ({
 		},
 		[profiles, updateProfiles],
 	)
+
 	const swapProfiles = useCallback(
 		(fromIndex: number, toIndex: number) => {
 			const newProfiles = [...profiles]
@@ -158,156 +137,15 @@ export const VirtualQuotaFallbackProvider = ({
 		[profiles],
 	)
 
-	return (
-		<>
-			<h3 className="text-lg font-medium mb-0">
-				<Trans i18nKey="kilocode:virtualProvider.title">Virtual Quota Fallback Settings</Trans>
-			</h3>
-			<div className="text-sm text-vscode-descriptionForeground mb-4">
-				<Trans i18nKey="kilocode:virtualProvider.description">
-					Configure a list of profiles each with their own limits. When one profiles limits are reached, the
-					next profile in the list will be used until none remain.
-				</Trans>
-			</div>
-
-			<div className="space-y-1">
-				{profiles.map((profile, index) => {
-					const usedProfileIds = getUsedProfileIds(index)
-					const availableForThisSlot = availableProfiles.filter(
-						(profile) => !usedProfileIds.includes(profile.id),
-					)
-
-					return (
-						<div key={index} className="border border-vscode-settings-sashBorder rounded-md p-2">
-							<div className="flex items-center justify-between mb-3">
-								<label className="block font-medium">
-									{index === 0
-										? t("kilocode:virtualProvider.primaryProfileLabel", { number: index + 1 })
-										: t("kilocode:virtualProvider.profileLabel", { number: index + 1 })}
-								</label>
-								<div className="flex items-center gap-1">
-									{/* Move Up Button */}
-									<VSCodeButton
-										appearance="icon"
-										onClick={() => moveProfileUp(index)}
-										disabled={index === 0}
-										title={t("kilocode:virtualProvider.moveProfileUp")}>
-										<ChevronUp size={16} />
-									</VSCodeButton>
-									{/* Move Down Button */}
-									<VSCodeButton
-										appearance="icon"
-										onClick={() => moveProfileDown(index)}
-										disabled={index === profiles.length - 1}
-										title={t("kilocode:virtualProvider.moveProfileDown")}>
-										<ChevronDown size={16} />
-									</VSCodeButton>
-									{/* Remove Button */}
-									{profiles.length > 1 && (
-										<VSCodeButton
-											appearance="icon"
-											onClick={() => removeProfile(index)}
-											title={t("kilocode:virtualProvider.removeProfile")}>
-											<TrashIcon />
-										</VSCodeButton>
-									)}
-								</div>
-							</div>
-
-							<Select
-								value={profile.profileId || ""}
-								onValueChange={(value) => handleProfileSelect(index, value)}
-								disabled={availableForThisSlot.length === 0}>
-								<SelectTrigger className="w-full">
-									<SelectValue placeholder={t("kilocode:virtualProvider.selectProfilePlaceholder")} />
-								</SelectTrigger>
-								<SelectContent>
-									{availableForThisSlot.map((profile) => (
-										<SelectItem key={profile.id} value={profile.id}>
-											{profile.name}
-										</SelectItem>
-									))}
-								</SelectContent>
-							</Select>
-
-							<VirtualLimitInputs profile={profile} index={index} onProfileChange={handleProfileChange} />
-						</div>
-					)
-				})}
-
-				<div className="flex justify-center p-4">
-					<VSCodeButton
-						appearance="secondary"
-						onClick={addProfile}
-						disabled={availableProfiles.length <= profiles.length}>
-						<PlusIcon className="mr-2" />
-						<Trans i18nKey="kilocode:virtualProvider.addProfile">Add Profile</Trans>
-					</VSCodeButton>
-				</div>
-
-				{availableProfiles.length === 0 ? (
-					<div className="text-sm text-vscode-descriptionForeground text-center p-4 border border-vscode-settings-sashBorder rounded-md">
-						<Trans i18nKey="kilocode:virtualProvider.noProfilesAvailable">
-							No profile profiles available. Please configure at least one non-virtual profile profile
-							first.
-						</Trans>
-					</div>
-				) : null}
-			</div>
-
-			<div className="p-4 border border-vscode-editorWarning-foreground rounded-md">
-				<div className="text-md font-semibold text-vscode-editorWarning-foreground">
-					<Trans i18nKey="kilocode:virtualProvider.dangerZoneTitle">Danger Zone</Trans>
-				</div>
-				<p className="text-sm text-vscode-descriptionForeground mt-1 mb-3">
-					<Trans i18nKey="kilocode:virtualProvider.dangerZoneDescription">
-						These actions are destructive and cannot be undone.
-					</Trans>
-				</p>
-				<VSCodeButton appearance="secondary" onClick={() => setIsAlertOpen(true)}>
-					<Trans i18nKey="kilocode:virtualProvider.clearUsageData">Clear Usage Data</Trans>
-				</VSCodeButton>
-			</div>
-
-			<AlertDialog open={isAlertOpen} onOpenChange={setIsAlertOpen}>
-				<AlertDialogContent>
-					<AlertDialogHeader>
-						<AlertDialogTitle>
-							<Trans i18nKey="kilocode:virtualProvider.confirmClearTitle">Are you sure?</Trans>
-						</AlertDialogTitle>
-						<AlertDialogDescription>
-							<Trans i18nKey="kilocode:virtualProvider.confirmClearDescription">
-								This will permanently delete all stored usage data for virtual profiles. This action
-								cannot be undone.
-							</Trans>
-						</AlertDialogDescription>
-					</AlertDialogHeader>
-					<AlertDialogFooter>
-						<AlertDialogCancel>
-							<Trans i18nKey="common:cancel">Cancel</Trans>
-						</AlertDialogCancel>
-						<AlertDialogAction onClick={handleClearUsageData}>
-							<Trans i18nKey="common:confirm">Confirm</Trans>
-						</AlertDialogAction>
-					</AlertDialogFooter>
-				</AlertDialogContent>
-			</AlertDialog>
-		</>
-	)
-}
-
-const VirtualLimitInputs = ({ profile, index, onProfileChange }: LimitInputsProps) => {
-	const [usage, setUsage] = useState<UsageResultByDuration | null>(null)
-
+	// Handle usage data fetching
 	useEffect(() => {
-		if (profile.profileId) {
-			vscode.postMessage({ type: "getUsageData", text: profile.profileId })
-		}
-
 		const handleMessage = (event: MessageEvent) => {
 			const message = event.data
-			if (message.type === "usageDataResponse" && message.text === profile.profileId) {
-				setUsage(message.values)
+			if (message.type === "usageDataResponse" && message.text) {
+				setUsageData((prev) => ({
+					...prev,
+					[message.text]: message.values,
+				}))
 			}
 		}
 
@@ -315,132 +153,32 @@ const VirtualLimitInputs = ({ profile, index, onProfileChange }: LimitInputsProp
 		return () => {
 			window.removeEventListener("message", handleMessage)
 		}
-	}, [profile.profileId])
+	}, [])
 
-	const handleLimitChange = useCallback(
-		(limitKey: keyof NonNullable<VirtualQuotaFallbackProviderData["profileLimits"]>) => (event: unknown) => {
-			const value = inputEventTransform(event)
-			const updatedProfile = {
-				...profile,
-				profileLimits: {
-					...profile.profileLimits,
-					[limitKey]: value === "" ? undefined : Number(value),
-				},
+	// Fetch usage data for all profiles
+	useEffect(() => {
+		profiles.forEach((profile) => {
+			if (profile.profileId) {
+				vscode.postMessage({ type: "getUsageData", text: profile.profileId })
 			}
-			onProfileChange(index, updatedProfile)
-		},
-		[profile, index, onProfileChange],
-	)
-
-	const renderProgressBarForLimit = useCallback(
-		(
-			limitType: keyof NonNullable<VirtualQuotaFallbackProviderData["profileLimits"]>,
-			usageKey: "tokens" | "requests",
-			duration: "minute" | "hour" | "day",
-		) => {
-			const limit = profile.profileLimits?.[limitType]
-			if (!limit || !usage) return null
-
-			const current = usage[duration][usageKey]
-
-			return (
-				<div className="mt-1">
-					<LabeledProgress label="" currentValue={current} limitValue={limit} />
-				</div>
-			)
-		},
-		[profile.profileLimits, usage],
-	)
-
-	if (!profile.profileId) {
-		return null
-	}
+		})
+	}, [profiles])
 
 	return (
-		<div className="space-y-4 p-2 rounded-md mt-2">
-			{/* Tokens Row */}
-			<div>
-				<label className="block text-sm font-medium mb-2">
-					<Trans i18nKey="kilocode:virtualProvider.tokensLabel">Tokens</Trans>
-				</label>
-				<div className="grid grid-cols-3 gap-x-4">
-					<div>
-						<label className="block text-xs text-vscode-descriptionForeground mb-1">
-							<Trans i18nKey="kilocode:virtualProvider.perMinute">Per Minute</Trans>
-						</label>
-						<VSCodeTextField
-							value={profile.profileLimits?.tokensPerMinute?.toString() ?? ""}
-							onInput={handleLimitChange("tokensPerMinute")}
-							className="w-full"
-						/>
-						{renderProgressBarForLimit("tokensPerMinute", "tokens", "minute")}
-					</div>
-					<div>
-						<label className="block text-xs text-vscode-descriptionForeground mb-1">
-							<Trans i18nKey="kilocode:virtualProvider.perHour">Per Hour</Trans>
-						</label>
-						<VSCodeTextField
-							value={profile.profileLimits?.tokensPerHour?.toString() ?? ""}
-							onInput={handleLimitChange("tokensPerHour")}
-							className="w-full"
-						/>
-						{renderProgressBarForLimit("tokensPerHour", "tokens", "hour")}
-					</div>
-					<div>
-						<label className="block text-xs text-vscode-descriptionForeground mb-1">
-							<Trans i18nKey="kilocode:virtualProvider.perDay">Per Day</Trans>
-						</label>
-						<VSCodeTextField
-							value={profile.profileLimits?.tokensPerDay?.toString() ?? ""}
-							onInput={handleLimitChange("tokensPerDay")}
-							className="w-full"
-						/>
-						{renderProgressBarForLimit("tokensPerDay", "tokens", "day")}
-					</div>
-				</div>
-			</div>
-
-			{/* Requests Row */}
-			<div>
-				<label className="block text-sm font-medium mb-2">
-					<Trans i18nKey="kilocode:virtualProvider.requestsLabel">Requests</Trans>
-				</label>
-				<div className="grid grid-cols-3 gap-x-4">
-					<div>
-						<label className="block text-xs text-vscode-descriptionForeground mb-1">
-							<Trans i18nKey="kilocode:virtualProvider.perMinute">Per Minute</Trans>
-						</label>
-						<VSCodeTextField
-							value={profile.profileLimits?.requestsPerMinute?.toString() ?? ""}
-							onInput={handleLimitChange("requestsPerMinute")}
-							className="w-full"
-						/>
-						{renderProgressBarForLimit("requestsPerMinute", "requests", "minute")}
-					</div>
-					<div>
-						<label className="block text-xs text-vscode-descriptionForeground mb-1">
-							<Trans i18nKey="kilocode:virtualProvider.perHour">Per Hour</Trans>
-						</label>
-						<VSCodeTextField
-							value={profile.profileLimits?.requestsPerHour?.toString() ?? ""}
-							onInput={handleLimitChange("requestsPerHour")}
-							className="w-full"
-						/>
-						{renderProgressBarForLimit("requestsPerHour", "requests", "hour")}
-					</div>
-					<div>
-						<label className="block text-xs text-vscode-descriptionForeground mb-1">
-							<Trans i18nKey="kilocode:virtualProvider.perDay">Per Day</Trans>
-						</label>
-						<VSCodeTextField
-							value={profile.profileLimits?.requestsPerDay?.toString() ?? ""}
-							onInput={handleLimitChange("requestsPerDay")}
-							className="w-full"
-						/>
-						{renderProgressBarForLimit("requestsPerDay", "requests", "day")}
-					</div>
-				</div>
-			</div>
-		</div>
+		<VirtualQuotaFallbackProviderPresentation
+			profiles={profiles}
+			availableProfiles={availableProfiles}
+			isAlertOpen={isAlertOpen}
+			usageData={usageData}
+			onProfileChange={handleProfileChange}
+			onProfileSelect={handleProfileSelect}
+			onAddProfile={addProfile}
+			onRemoveProfile={removeProfile}
+			onMoveProfileUp={moveProfileUp}
+			onMoveProfileDown={moveProfileDown}
+			onClearUsageData={handleClearUsageData}
+			onSetIsAlertOpen={setIsAlertOpen}
+			getUsedProfileIds={getUsedProfileIds}
+		/>
 	)
 }
