+++ webview-ui/src/components/ui/labeled-progress.tsx
@@ -12,7 +12,7 @@ export const LabeledProgress = ({ label, currentValue, limitValue, className }: 
 
 	return (
 		<div className={className}>
-			<Progress value={percentage} className="[&>div]:bg-vscode-button-background mt-2" />
+			<Progress value={percentage} className="[&>div]:bg-vscode-button-background mt-1 h-1" />
 			<div className="text-xs text-vscode-descriptionForeground mb-1 flex gap-2 justify-between">
 				<span className="whitespace-nowrap align-end">{label}</span>
 				<span className="whitespace-nowrap flex text-[10px] items-end">
